+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 ## [Unreleased]
 
+## [0.1.0] - 2021-03-17
+
+-   Creates mlflow experiment on pipeline start if it does not exist
+-   Temporary persistent volumes generation and removal for pipelines
+-   `upload-pipeline` command that generates DAG to Airflow directory 
+-   `schedule` command that generates DAG to Airflow directory with the given schedule
+-    `run-once` command that uploads pipeline to Airflow and creates the DAG run
 ## [0.0.4] - 2021-03-10
+
+### Fixed
+
+-   Added DAG template to package manifest
 
 ## [0.0.3] - 2021-03-09
 
@@ -16,7 +27,9 @@
 
 -   Initial implementation of `kedro_airflow_k8s` plugin stub. 
 
-[Unreleased]: https://github.com/getindata/kedro-airflow-k8s/compare/0.0.4...HEAD
+[Unreleased]: https://github.com/getindata/kedro-airflow-k8s/compare/0.1.0...HEAD
+
+[0.1.0]: https://github.com/getindata/kedro-airflow-k8s/compare/0.0.4...0.1.0
 
 [0.0.4]: https://github.com/getindata/kedro-airflow-k8s/compare/0.0.3...0.0.4